+++ NEWS.md
@@ -1,3 +1,8 @@
+# tessistream 0.5.0
+
+* incremental `stream()` building from all existing streams
+* optimizations to `stream_customer_history` and `setnafill_group`
+
 # tessistream 0.4.1
 
 * email_stream now can be built incrementally, for more efficiency